requirements.txt
Flask==3.0.3
prometheus-client==0.20.0
flask-limiter==3.7.0
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
open_clip_torch==2.24.0
boto3==1.35.1